requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
efficientnet-pytorch==0.7.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.1
haversine==2.8.1
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.4
imageio==2.33.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
munch==4.0.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.17.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
Pillow==9.3.0
pip==23.3.1
pretrainedmodels==0.7.4
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
PyYAML==6.0.1
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
segmentation-models-pytorch==0.3.3
setuptools==68.2.2
six==1.16.0
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.2
torchmetrics==1.3.0.post0
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
